Bring some stuff more inline with standard C++ Library.  However still have to worry about fd opens for spectrum IO

+++ SpecTcl/XamineEventHandler.cpp
@@ -116,7 +116,7 @@ void CXamineEventHandler::operator()(int mask)
 
       m_nFd = m_pHistogrammer->getDisplayer()->GetEventFd();
       Tcl_Close(m_pInterp->getInterpreter(), m_SocketChannel);
-      m_SocketChannel = Tcl_MakeTcpClientChannel((ClientData)m_nFd);
+      m_SocketChannel = Tcl_MakeTcpClientChannel((ClientData)(long)m_nFd);
       Set(TK_READABLE);
       cerr << "\n";
     }
